+++ pkg/commands/sso/sso_test.go
@@ -9,6 +9,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/fastly/go-fastly/v9/fastly"
+
 	"github.com/fastly/cli/pkg/api"
 	"github.com/fastly/cli/pkg/app"
 	"github.com/fastly/cli/pkg/auth"
@@ -85,11 +87,12 @@ func TestSSO(t *testing.T) {
 			TestScenario: testutil.TestScenario{
 				Args: args("sso"),
 				WantOutputs: []string{
-					"We're going to authenticate the 'user' profile.",
+					"We're going to authenticate the 'user' profile",
 					"We need to open your browser to authenticate you.",
 					"Session token (persisted to your local configuration): 123",
 				},
 			},
+			HTTPClient: testutil.CurrentCustomerClient(testutil.CurrentCustomerResponse),
 			AuthResult: &auth.AuthorizationResult{
 				SessionToken: "123",
 			},
@@ -105,11 +108,12 @@ func TestSSO(t *testing.T) {
 			TestScenario: testutil.TestScenario{
 				Args: args("sso test_user"),
 				WantOutputs: []string{
-					"We're going to authenticate the 'test_user' profile.",
+					"We're going to authenticate the 'test_user' profile",
 					"We need to open your browser to authenticate you.",
 					"Session token (persisted to your local configuration): 123",
 				},
 			},
+			HTTPClient: testutil.CurrentCustomerClient(testutil.CurrentCustomerResponse),
 			AuthResult: &auth.AuthorizationResult{
 				SessionToken: "123",
 			},
@@ -138,11 +142,29 @@ func TestSSO(t *testing.T) {
 		// Otherwise no OAuth flow is happening here.
 		{
 			TestScenario: testutil.TestScenario{
-				Args: args("whoami"),
+				API: mock.API{
+					AllDatacentersFn: func() ([]fastly.Datacenter, error) {
+						return []fastly.Datacenter{
+							{
+								Name:   fastly.ToPointer("Foobar"),
+								Code:   fastly.ToPointer("FBR"),
+								Group:  fastly.ToPointer("Bar"),
+								Shield: fastly.ToPointer("Baz"),
+								Coordinates: &fastly.Coordinates{
+									Latitude:   fastly.ToPointer(float64(1)),
+									Longtitude: fastly.ToPointer(float64(2)),
+									X:          fastly.ToPointer(float64(3)),
+									Y:          fastly.ToPointer(float64(4)),
+								},
+							},
+						}, nil
+					},
+				},
+				Args: args("pops"),
 				WantOutputs: []string{
 					// FIXME: Put back messaging once SSO is GA.
 					// "is not a Fastly SSO (Single Sign-On) generated token",
-					"Alice Programmer <alice@example.com>",
+					"{Latitude:1 Longtitude:2 X:3 Y:4}",
 				},
 			},
 			ConfigFile: &config.File{
@@ -157,7 +179,7 @@ func TestSSO(t *testing.T) {
 			ExpectedConfigProfile: &config.Profile{
 				Token: "mock-token",
 			},
-			HTTPClient: testutil.WhoamiVerifyClient(testutil.WhoamiBasicResponse),
+			HTTPClient: testutil.CurrentCustomerClient(testutil.CurrentCustomerResponse),
 		},
 		// 7. Success processing `whoami` command.
 		// We set an SSO token that has expired.
@@ -168,7 +190,7 @@ func TestSSO(t *testing.T) {
 			TestScenario: testutil.TestScenario{
 				Args:           args("whoami"),
 				WantOutput:     "Your access token has expired and so has your refresh token.",
-				DontWantOutput: "Alice Programmer <alice@example.com>",
+				DontWantOutput: "{Latitude:1 Longtitude:2 X:3 Y:4}",
 			},
 			ConfigFile: &config.File{
 				Profiles: config.Profiles{
@@ -183,19 +205,37 @@ func TestSSO(t *testing.T) {
 			ExpectedConfigProfile: &config.Profile{
 				Token: "mock-token",
 			},
-			HTTPClient: testutil.WhoamiVerifyClient(testutil.WhoamiBasicResponse),
+			HTTPClient: testutil.CurrentCustomerClient(testutil.CurrentCustomerResponse),
 		},
 		// 8. Success processing OAuth flow via `whoami` command
 		// We set an SSO token that has expired.
 		// This allows us to validate the output messages.
 		{
 			TestScenario: testutil.TestScenario{
-				Args: args("whoami"),
+				API: mock.API{
+					AllDatacentersFn: func() ([]fastly.Datacenter, error) {
+						return []fastly.Datacenter{
+							{
+								Name:   fastly.ToPointer("Foobar"),
+								Code:   fastly.ToPointer("FBR"),
+								Group:  fastly.ToPointer("Bar"),
+								Shield: fastly.ToPointer("Baz"),
+								Coordinates: &fastly.Coordinates{
+									Latitude:   fastly.ToPointer(float64(1)),
+									Longtitude: fastly.ToPointer(float64(2)),
+									X:          fastly.ToPointer(float64(3)),
+									Y:          fastly.ToPointer(float64(4)),
+								},
+							},
+						}, nil
+					},
+				},
+				Args: args("pops"),
 				WantOutputs: []string{
 					"Your access token has expired and so has your refresh token.",
 					"Starting a local server to handle the authentication flow.",
 					"Session token (persisted to your local configuration): 123",
-					"Alice Programmer <alice@example.com>",
+					"{Latitude:1 Longtitude:2 X:3 Y:4}",
 				},
 			},
 			AuthResult: &auth.AuthorizationResult{
@@ -214,7 +254,7 @@ func TestSSO(t *testing.T) {
 			ExpectedConfigProfile: &config.Profile{
 				Token: "123",
 			},
-			HTTPClient: testutil.WhoamiVerifyClient(testutil.WhoamiBasicResponse),
+			HTTPClient: testutil.CurrentCustomerClient(testutil.CurrentCustomerResponse),
 			Stdin: []string{
 				"Y", // when prompted to open a web browser to start authentication
 			},