+++ pkg/errors/errors.go
@@ -147,7 +147,7 @@ var ErrNoSTDINData = RemediationError{
 
 // ErrInvalidKVCombo means the user omitted either the key or value flag.
 var ErrInvalidKVCombo = RemediationError{
-	Inner:       fmt.Errorf("--key-name and --value are required"),
+	Inner:       fmt.Errorf("--key and --value are required"),
 	Remediation: "Please add both flags or alternatively use either --stdin or --file.",
 }